# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NoteMarkdown metadata API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "ClientDiagnosticAccepted": {
        "type": "object",
        "properties": {
          "reportId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "reportId"
        ],
        "additionalProperties": false
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "ClientDiagnosticReport": {
        "type": "object",
        "properties": {
          "reportId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "integer",
            "minimum": 0
          },
          "trigger": {
            "type": "string",
            "enum": [
              "workspace-error",
              "unhandled-error",
              "unhandled-rejection",
              "slow-operation"
            ]
          },
          "buildMode": {
            "type": "string",
            "maxLength": 40
          },
          "pageState": {
            "type": "object",
            "properties": {
              "online": {
                "type": "boolean"
              },
              "visibility": {
                "type": "string",
                "enum": [
                  "visible",
                  "hidden",
                  "prerender",
                  "unloaded",
                  "unknown"
                ]
              },
              "providerType": {
                "type": "string",
                "enum": [
                  "drive",
                  "local",
                  "none"
                ]
              },
              "isOpening": {
                "type": "boolean"
              },
              "isIndexing": {
                "type": "boolean"
              },
              "entryCount": {
                "type": "integer",
                "minimum": 0
              },
              "tabCount": {
                "type": "integer",
                "minimum": 0
              },
              "saveStates": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            },
            "required": [
              "online",
              "visibility",
              "providerType",
              "isOpening",
              "isIndexing",
              "entryCount",
              "tabCount",
              "saveStates"
            ],
            "additionalProperties": false
          },
          "failure": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "maxLength": 100
              },
              "code": {
                "type": "string",
                "maxLength": 100
              },
              "status": {
                "type": "integer",
                "minimum": 0,
                "maximum": 599
              },
              "apiCode": {
                "type": "string",
                "maxLength": 100
              },
              "stackFrames": {
                "type": "array",
                "items": {
                  "type": "string",
                  "maxLength": 300
                },
                "maxItems": 20
              },
              "causeNames": {
                "type": "array",
                "items": {
                  "type": "string",
                  "maxLength": 100
                },
                "maxItems": 8
              }
            },
            "required": [
              "name",
              "stackFrames",
              "causeNames"
            ],
            "additionalProperties": false
          },
          "metrics": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientDiagnosticEvent"
            },
            "maxItems": 300
          }
        },
        "required": [
          "reportId",
          "createdAt",
          "trigger",
          "buildMode",
          "pageState",
          "failure",
          "metrics",
          "events"
        ],
        "additionalProperties": false
      },
...
      "Me": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "connectedAccounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectedAccount"
            }
          }
        },
        "required": [
          "user",
          "connectedAccounts"
        ]
      },
...
      "Success": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
      "DriveToken": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "integer"
          }
        },
        "required": [
          "accessToken",
          "expiresAt"
        ]
      },
      "TokenRequest": {
        "type": "object",
        "properties": {
          "connectedAccountId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "connectedAccountId"
        ],
        "additionalProperties": false
      },
      "WorkspaceList": {
        "type": "object",
        "properties": {
          "workspaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DriveWorkspace"
            }
          }
        },
        "required": [
          "workspaces"
        ]
      },
...
    },
...
  },
  "paths": {
    "/api/v1/diagnostics/client-errors": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDiagnosticReport"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Diagnostic report accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDiagnosticAccepted"
                }
              }
            }
          },
          "404": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/start": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "returnTo",
            "in": "query"
          }
        ],
        "responses": {
          "302": {
            "description": "Google OAuth redirect"
          },
          "400": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/callback": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "state",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "code",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "error",
            "in": "query"
          }
        ],
        "responses": {
          "302": {
            "description": "Application redirect"
          },
          "400": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "responses": {
          "200": {
            "description": "Current identity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "401": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/session/logout": {
      "post": {
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drive/token": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Short-lived browser-to-Drive token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveToken"
                }
              }
            }
          },
          "401": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drive/workspaces": {
      "get": {
        "responses": {
          "200": {
            "description": "Linked Drive folders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceList"
                }
              }
            }
          },
          "401": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspace"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Linked Drive folder",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriveWorkspace"
                }
              }
            }
          },
          "401": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drive/workspaces/{id}": {
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Reference removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/preferences": {
      "get": {
        "responses": {
          "200": {
            "description": "Global preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preference"
                }
              }
            }
          },
          "401": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preference"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preference"
                }
              }
            }
          },
          "401": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/connected-accounts/{id}": {
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider disconnected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/account": {
      "delete": {
        "responses": {
          "200": {
            "description": "Account metadata deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Stable API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
...
}